+++ pytmbot/keyboards/keyboards.py
@@ -119,7 +119,7 @@ class Keyboards:
         ):
             keyboard = ReplyKeyboardMarkup(
                 resize_keyboard=True,
-                one_time_keyboard=True,
+                one_time_keyboard=False,
                 selective=True,
                 is_persistent=True,
             )